# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API RESTFUL Laravel 5.8",
        "description": "API documentation project",
        "contact": {
            "email": "carlos.clayton@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories/trashed": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "List of trashed categories",
                "description": "Return a list of trashed categories",
                "operationId": "ApiVue\\Http\\Controllers\\CategoriesController::trashed",
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
        "/api/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "List of clients",
                "description": "Return a list of clients",
                "operationId": "ApiVue\\Http\\Controllers\\ClientsController::index",
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List of orders",
                "description": "Return a list of orders",
                "operationId": "ApiVue\\Http\\Controllers\\OrdersController::index",
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Store a order",
                "description": "Return message",
                "operationId": "ApiVue\\Http\\Controllers\\OrdersController::store",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store a order"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Show a order",
                "description": "Return a order",
                "operationId": "getOrderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update a order",
                "description": "Return message",
                "operationId": "getOrderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Store a order"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Orders"
                ],
                "summary": "Delete a order",
                "description": "Delete a order",
                "operationId": "getOrderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of order",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/orders/trashed": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "List of trashed orders",
                "description": "Return a list of trashed orders",
                "operationId": "ApiVue\\Http\\Controllers\\OrdersController::trashed",
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/orders/restore/{id}": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Restore a order",
                "description": "Restore a order",
                "operationId": "getOrderById",
                "responses": {
                    "200": {
                        "description": "Store categories"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
        "/api/products/trashed": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "List of trashed products",
                "description": "Return a list of trashed products",
                "operationId": "ApiVue\\Http\\Controllers\\ProductsController::trashed",
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
        "/api/users/trashed": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List of trashed users",
                "description": "Return a list of users",
                "operationId": "ApiVue\\Http\\Controllers\\UsersController::trashed",
                "responses": {
                    "200": {
                        "description": "An json"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Use a global client_id / client_secret and your username / password combo to obtain a token",
                "name": "Authorization",
                "in": "header",
                "scheme": "http"
            }
        }
    }
}